% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyover.R
\docType{package}
\name{dplyover-package}
\alias{dplyover}
\alias{dplyover-package}
\title{dplyover: Create Columns by Looping over Strings in 'dplyr'}
\description{
To learn more about dplyover, start with the vignette:
\code{browseVignettes(package = "dplyover")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TimTeaFan/dplyover}
  \item Report bugs at \url{https://github.com/TimTeaFan/dplyover/issues}
}

}
\author{
\strong{Maintainer}: Tim Tiefenbach \email{mailme@tim-tiefenbach.de} (\href{https://orcid.org/0000-0001-9443-2434}{ORCID})

}
