% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_across_family.R
\name{over_across_family}
\alias{over_across_family}
\title{The over-across function family}
\description{
\code{dplyover} extends \code{dplyr}'s functionality by building a function family
around \code{dplyr::across()}.

The goal of this \strong{over-across function family} is to provide a concise and
uniform syntax which can be used to create columns by applying functions to
vectors and / or sets of columns in dplyr. Ideally, this will improve our
mental model so that it is easier to tackle problems where the solution is
based on creating new columns.

The functions in the over-apply function family create columns by applying
one or several functions to:
\subsection{basic functions}{
\itemize{
\item \code{\link[dplyr:across]{dplyr::across()}}: a set of columns
\item \code{\link[=over]{over()}}: a vector (list or atomic vector)
}
}

\subsection{variants}{
\itemize{
\item \code{\link[=over2]{over2()}} two vectors of the same length (pairwise)
\item \code{\link[=over2x]{over2x()}} two vectors (nested)
\item \code{\link[=across2]{across2()}} two sets of columns (pairwise)
\item \code{\link[=across2x]{across2x()}} two sets of columns (nested)
\item \code{\link[=crossoverx]{crossoverx()}} a set of columns and a vector (nested)
}
}
}
