% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.R
\name{.}
\alias{.}
\title{\code{over()}'s genuine forcing function}
\usage{
.(x)
}
\arguments{
\item{.pattern}{Pattern to look for.}

\item{.vars}{A character vector with variables names. When used inside \code{over}
all column names of the underlying data are automatically supplied to \code{.vars}.
This argument is useful when testing the functionality outside the context of
\code{over()}.}

\item{.select}{Pattern to further select and narrow down the variable names
provided in \code{.vars}. When this argument is provided the variables names in
\code{.vars} will be narrowed down to those who match the pattern specified in
\code{.select}.}
}
\value{
A character vector.
}
\description{
These functions are \link[=selection_helpers]{selection helpers}. They are intended
to be used inside \code{over()} to extract parts or patterns of the column names of
the underlying data.
\itemize{
\item \code{\link[=cut_names]{cut_names()}} selects strings by removing (cutting off) the specified \code{.pattern}.
This functionality resembles \code{stringr::str_remove_all()}.
\item \code{\link[=extract_names]{extract_names()}} selects strings by extracting the specified \code{.pattern}.
This functionality resembles \code{stringr::str_extract()}.
}
}
\section{Examples}{

}

